Plugin update auth_outage.

+++ auth/outage/classes/local/outagelib.php
@@ -14,20 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-/**
- * outagelib class.
- *
- * @package    auth_outage
- * @author     Daniel Thee Roperto <daniel.roperto@catalyst-au.net>
- * @copyright  2016 Catalyst IT
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-
 namespace auth_outage\local;
 
 use auth_outage\dml\outagedb;
 use auth_outage\local\controllers\maintenance_static_page;
-use auth_outage\output\renderer;
 use coding_exception;
 use curl;
 use Exception;
@@ -36,8 +26,7 @@ use invalid_parameter_exception;
 use stdClass;
 
 defined('MOODLE_INTERNAL') || die();
-
-require_once(__DIR__.'/../../lib.php');
+require_once(__DIR__ . '/../../lib.php');
 
 /**
  * outagelib class.
@@ -48,7 +37,6 @@ require_once(__DIR__.'/../../lib.php');
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class outagelib {
-
     /** Outage start. */
     const OUTAGE_START = '<!-- OUTAGESTART -->';
 
@@ -146,7 +134,7 @@ class outagelib {
             $renderer = $PAGE->get_renderer('auth_outage');
             return $renderer->render_warningbar($active, $time, false, $preview);
         } catch (Exception $e) {
-            debugging('Exception occured while injecting our code: '.$e->getMessage());
+            debugging('Exception occured while injecting our code: ' . $e->getMessage());
             debugging($e->getTraceAsString(), DEBUG_DEVELOPER);
         }
     }
@@ -185,7 +173,6 @@ class outagelib {
             'allowedips'               => '',
             'css'                      => '',
             'default_time'             => '',
-            'default_autostart'        => '0',
             'default_duration'         => (string)(60 * 60),
             'default_warning_duration' => (string)(60 * 60),
             'default_title'            => get_string('defaulttitlevalue', 'auth_outage'),
@@ -206,9 +193,6 @@ class outagelib {
         $outage = outagedb::get_ongoing();
         if (is_null($outage)) {
             $outage = outagedb::get_next_starting();
-            $ongoingoutage = false;
-        } else {
-            $ongoingoutage = true;
         }
 
         // Set json formatted outage string to cache.
@@ -216,9 +200,6 @@ class outagelib {
 
         maintenance_static_page::create_from_outage($outage)->generate();
         self::update_climaintenance_code($outage);
-        if (!$ongoingoutage || $reenablemaint || is_null($outage)) {
-            self::update_maintenance_later($outage);
-        }
     }
 
     /**
@@ -230,25 +211,6 @@ class outagelib {
         return (!empty($result['contents']));
     }
 
-    /**
-     * Calls Moodle API - set_maintenance_later() to set when the next outage starts.
-     * @param outage|null $outage Outage or null if no scheduled outage.
-     */
-    private static function update_maintenance_later($outage) {
-        if (is_null($outage) || !$outage->autostart) {
-            unset_config('maintenance_later');
-        } else {
-            $message = get_config('moodle', 'maintenance_message');
-            if ($message) {
-                debugging('Disabling $CFG->maintenance_message to allow our template page to take place.');
-                debugging('Previous value: '.$message);
-                // We cannot do much if forced config, but the logs will show the error.
-                unset_config('maintenance_message');
-            }
-            set_config('maintenance_later', $outage->starttime);
-            self::maintenance_config_log($outage);
-        }
-    }
 
     /**
      * Checks if we should try to inject an warning bar.
@@ -264,7 +226,7 @@ class outagelib {
         self::$injectcalled = true;
 
         // Do not inject into admin/settings.php.
-        if ($_SERVER['SCRIPT_NAME'] == '/'.$CFG->admin.'/settings.php') {
+        if ($_SERVER['SCRIPT_NAME'] == '/' . $CFG->admin . '/settings.php') {
             if (optional_param('section', '', PARAM_RAW) === 'additionalhtml') {
                 return false;
             }
@@ -277,7 +239,7 @@ class outagelib {
 
         // Used to test the try block in case of errors.
         if (PHPUNIT_TEST && optional_param('auth_outage_break_code', false, PARAM_INT)) {
-            (new stdClass())->invalidfield; // Triggers an exception.
+            throw new coding_exception('Exception triggered');
         }
 
         // Nothing preventing the injection.
@@ -290,57 +252,102 @@ class outagelib {
      * @param int    $starttime  Outage start time.
      * @param int    $stoptime   Outage stop time.
      * @param string $allowedips List of IPs allowed.
+     * @param string|null $accesskey access key, or null if no access key set.
+     * @param string|null $metadata metadata to be added to the outage headers, or null if none.
      *
      * @return string
      * @throws invalid_parameter_exception
      */
-    public static function create_climaintenancephp_code($starttime, $stoptime, $allowedips) {
+    public static function create_climaintenancephp_code($starttime, $stoptime, $allowedips, $accesskey = null, $metadata = null) {
+        global $CFG;
         if (!is_int($starttime) || !is_int($stoptime)) {
             throw new invalid_parameter_exception('Make sure $startime and $stoptime are integers.');
-        }
-        if (!is_string($allowedips) || (trim($allowedips) == '')) {
-            throw new invalid_parameter_exception('$allowedips must be a valid string.');
         }
         // I know Moodle validation would clean up this field, but just in case, let's ensure no
         // single-quotes (and double for the sake of it) are present otherwise it would break the code.
         $allowedips = addslashes($allowedips);
 
+        $cookiesecure = is_moodle_cookie_secure();
+
+        // Since Moodle 4.3 cookiehttponly is default to true and this CFG is not set.
+        // so if not set, default to true.
+        $cookiehttponly = isset($CFG->cookiehttponly) ? (bool) $CFG->cookiehttponly : true;
+
         $code = <<<'EOT'
 <?php
 if ((time() >= {{STARTTIME}}) && (time() < {{STOPTIME}})) {
-    define('MOODLE_INTERNAL', true);
+    if (!defined('MOODLE_INTERNAL')) {
+        define('MOODLE_INTERNAL', true);
+    }
     require_once($CFG->dirroot.'/lib/moodlelib.php');
     if (file_exists($CFG->dirroot.'/lib/classes/ip_utils.php')) {
         require_once($CFG->dirroot.'/lib/classes/ip_utils.php');
     }
-    if (!remoteip_in_list('{{ALLOWEDIPS}}')) {
-        header($_SERVER['SERVER_PROTOCOL'] . ' 503 Moodle under maintenance');
-        header('Status: 503 Moodle under maintenance');
-        header('Retry-After: 300');
-        header('Content-type: text/html; charset=utf-8');
-        header('X-UA-Compatible: IE=edge');
-        header('Cache-Control: no-store, no-cache, must-revalidate');
-        header('Cache-Control: post-check=0, pre-check=0', false);
-        header('Pragma: no-cache');
-        header('Expires: Mon, 20 Aug 1969 09:23:00 GMT');
-        header('Last-Modified: ' . gmdate('D, d M Y H:i:s') . ' GMT');
-        header('Accept-Ranges: none');
-        header('X-Moodle-Maintenance: manager');
-        if ((defined('AJAX_SCRIPT') && AJAX_SCRIPT) || (defined('WS_SERVER') && WS_SERVER)) {
+    // Put access key as a cookie if given. This stops the need to put it as a url param on every request.
+    $urlaccesskey = optional_param('accesskey', null, PARAM_TEXT);
+    $isphpunit = defined('PHPUNIT_TEST');
+
+    if (!empty($urlaccesskey) && !$isphpunit) {
+        setcookie('auth_outage_accesskey', $urlaccesskey, time() + 86400, '/', '', {{COOKIESECURE}}, {{COOKIEHTTPONLY}});
+    }
+
+    // Use url access key if given, else the cookie, else null.
+    $useraccesskey = $urlaccesskey ?: $_COOKIE['auth_outage_accesskey'] ?? null;
+
+    $ipblocked = !remoteip_in_list('{{ALLOWEDIPS}}');
+    $accesskeyblocked = $useraccesskey != '{{ACCESSKEY}}';
+    $allowed = ({{USEACCESSKEY}} && !$accesskeyblocked) || ({{USEALLOWEDIPS}} && !$ipblocked);
+
+    if (!$allowed) {
+        if (!$isphpunit) {
+            header($_SERVER['SERVER_PROTOCOL'] . ' 503 Moodle under maintenance');
+            header('Status: 503 Moodle under maintenance');
+            header('Retry-After: 300');
+            header('Content-type: text/html; charset=utf-8');
+            header('X-UA-Compatible: IE=edge');
+            header('Cache-Control: no-store, no-cache, must-revalidate');
+            header('Cache-Control: post-check=0, pre-check=0', false);
+            header('Pragma: no-cache');
+            header('Expires: Mon, 20 Aug 1969 09:23:00 GMT');
+            header('Last-Modified: ' . gmdate('D, d M Y H:i:s') . ' GMT');
+            header('Accept-Ranges: none');
+            header('X-Moodle-Maintenance: manager');
+            if (!empty({{METADATA}})) {
+                header('X-Outage-Metadata: ' . {{METADATA}});
+            }
+            header('X-Outage-StartTime: ' . '{{STARTTIME}}');
+            header('X-Outage-EndTime: ' . '{{STOPTIME}}');
+        }
+
+        if (!$isphpunit && ((defined('AJAX_SCRIPT') && AJAX_SCRIPT) || (defined('WS_SERVER') && WS_SERVER))) {
             exit(0);
         }
-        echo '<!-- Blocked by ip, your ip: '.getremoteaddr('n/a').' -->';
-        if (file_exists($CFG->dataroot.'/climaintenance.template.html')) {
-            require($CFG->dataroot.'/climaintenance.template.html');
-            exit(0);
+
+        if ({{USEALLOWEDIPS}} && $ipblocked) {
+            echo '<!-- auth_outage blocked your ip: '.getremoteaddr('n/a').' -->';
         }
-        // The file above should always exist, but just in case...
-        die('We are currently under maintentance, please try again later.');
+
+        if ({{USEACCESSKEY}} && $accesskeyblocked) {
+            echo '<!-- auth_outage blocked by missing or incorrect access key, access key given: '. $useraccesskey .' -->';
+        }
+
+        if (!$isphpunit) {
+            if (file_exists($CFG->dataroot.'/climaintenance.template.html')) {
+                require($CFG->dataroot.'/climaintenance.template.html');
+                exit(0);
+            }
+            // The file above should always exist, but just in case...
+            die('We are currently under maintentance, please try again later.');
+        }
     }
 }
 EOT;
-        $search = ['{{STARTTIME}}', '{{STOPTIME}}', '{{ALLOWEDIPS}}', '{{YOURIP}}'];
-        $replace = [$starttime, $stoptime, $allowedips, getremoteaddr('n/a')];
+        $search = ['{{STARTTIME}}', '{{STOPTIME}}', '{{USEALLOWEDIPS}}', '{{ALLOWEDIPS}}', '{{USEACCESSKEY}}', '{{ACCESSKEY}}',
+            '{{YOURIP}}', '{{COOKIESECURE}}', '{{COOKIEHTTPONLY}}', '{{METADATA}}'];
+        // Note that var_export is required because (string) false == '', not 'false'.
+        $replace = [$starttime, $stoptime, var_export(!empty($allowedips), true), $allowedips, var_export(!empty($accesskey), true),
+            $accesskey, getremoteaddr('n/a'), var_export($cookiesecure, true),
+            var_export($cookiehttponly, true), var_export($metadata, true)];
         return str_replace($search, $replace, $code);
     }
 
@@ -354,7 +361,7 @@ EOT;
      */
     public static function update_climaintenance_code($outage) {
         global $CFG;
-        $file = $CFG->dataroot.'/climaintenance.php';
+        $file = $CFG->dataroot . '/climaintenance.php';
 
         if (!is_null($outage) && !($outage instanceof outage)) {
             throw new coding_exception('$outage must be null or an outage object.');
@@ -362,17 +369,20 @@ EOT;
 
         $config = self::get_config();
         $allowedips = trim($config->allowedips);
+        $accesskey = $outage->accesskey ?? null;
+        $metadata = $outage->metadata ?? null;
 
-        if (is_null($outage) || ($allowedips == '')) {
+        // If no outage, or allowed ips is null and access key is null (i.e. no blocking required).
+        if (is_null($outage) || ($allowedips == '' && empty($accesskey))) {
             if (file_exists($file)) {
                 unlink($file);
             }
         } else {
-            $code = self::create_climaintenancephp_code($outage->starttime, $outage->stoptime, $allowedips);
+            $code = self::create_climaintenancephp_code($outage->starttime, $outage->stoptime, $allowedips, $accesskey, $metadata);
 
             $dir = dirname($file);
             if (!file_exists($dir) || !is_dir($dir)) {
-                throw new file_exception('Directory must exists: '.$dir);
+                throw new file_exception('Directory must exists: ' . $dir);
             }
             file_put_contents($file, $code);
         }
@@ -388,8 +398,10 @@ EOT;
 
         $message = [];
 
-        if (trim(self::get_config()->allowedips) != ''
-                && (!isset($CFG->auth_outage_bootstrap_loaded) || !$CFG->auth_outage_bootstrap_loaded)) {
+        if (
+            trim(self::get_config()->allowedips) != ''
+                && (!isset($CFG->auth_outage_bootstrap_loaded) || !$CFG->auth_outage_bootstrap_loaded)
+        ) {
             $message[] = get_string('configurationwarning', 'auth_outage');
         }
